% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{crowdtangle_token}
\alias{crowdtangle_token}
\title{Obtain CT Token from system environment}
\usage{
Warning: Never store your API token in plain text or print it in public scripts.
}
\value{
A string with a CT API Token.
}
\description{
Allows to retrieve a CT API token that has been previously written to the
system environment.
}
\examples{
# Set CT Token (Once per session)

 # set_crowdtangle_token(option = 'enter')

# Load token for use with CT API calls

#crowdtangle_token()
}
