% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{links}
\alias{links}
\title{R Interface for CT Link Search}
\usage{
links(
  token,
  count = 100,
  endDate = ct_now(),
  includeHistory = NULL,
  link = NULL,
  includeSummary = "false",
  offset = 0,
  platforms = NULL,
  searchField = NULL,
  startDate = NULL,
  sortBy = "date",
  output_raw = TRUE
)
}
\arguments{
\item{token}{A valid CT API Token}

\item{count}{...}

\item{endDate}{...}

\item{includeHistory}{...}

\item{link}{A URL}

\item{includeSummary}{... (Defaults to "false")}

\item{offset}{... (Usually you want to leave this at 0 and paginate via dates only)}

\item{platforms}{...}

\item{searchField}{...}

\item{startDate}{...}

\item{sortBy}{...}

\item{output_raw}{A boolean signifying whether to output raw CT output or parse into df. Defaulting to \code{TRUE}.}
}
\description{
Searches for link shares by public pages, groups, profiles on FB/IG
}
\examples{
postlist <-
links(token = crowdtangle_token(),
      count = 1000,
      endDate = '2020-12-31',
      includeHistory = NULL,
      link = "rt.com",
      includeSummary= 'false',
      offset = 0,
      platforms = 'facebook',
      searchField = 'Include_query_strings',
      startDate = '2020-06-01T00:00:00',
      sortBy = 'date')
}
