% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posts.R
\name{posts}
\alias{posts}
\title{R Interface for}
\usage{
posts(
  token,
  accounts = NULL,
  brandedContent = "no_filter",
  count = 10,
  endDate = ct_now(),
  includeHistory = NULL,
  language = NULL,
  listIds = NULL,
  minInteractions = 0,
  offset = 0,
  pageAdminTopCountry = NULL,
  searchTerm = "",
  startDate = NULL,
  sortBy = "date",
  timeframe = NULL,
  types = NULL,
  verified = "no_filter",
  search10k = FALSE,
  boolean_allowed = FALSE,
  output_raw = TRUE,
  error_wrapper = T
)
}
\arguments{
\item{token}{A valid CT API Token}

\item{accounts}{...}

\item{brandedContent}{...}

\item{count}{...}

\item{endDate}{...}

\item{includeHistory}{...}

\item{language}{...}

\item{listIds}{...}

\item{offset}{...}

\item{pageAdminTopCountry}{...}

\item{searchTerm}{...}

\item{startDate}{...}

\item{sortBy}{...}

\item{timeframe}{...}

\item{types}{...}

\item{verified}{...}

\item{search10k}{A boolean signifying whether you have elavated access for higher count numbers.}

\item{boolean_allowed}{A boolean signifying whether you have boolean term search enabled.}

\item{output_raw}{A boolean signifying whether to output raw CT output or parse into df. Defaulting to \code{TRUE}.}

\item{platforms}{...}

\item{searchField}{...}

\item{weights}{Not yet implemented}
}
\description{
Pulls posts
}
\examples{
# Pull all posts from November 2020 by CNN containing the term "Trump"
postlist <- posts_search(token = crowdtangle_token(),
                    inAccountIds = "8323", #CNN example
                    searchTerm = "Trump",
                    startDate = "2020-11-01T00:00:00",
                    endDate = "2020-11-30T23:59:59",
                    count = 10000,
                    sortBy = 'date',
                    platforms = 'facebook',
                    search10k = T,
                    boolean_allowed = T,
                    output_raw = TRUE)
}
