% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_crowdtangle_token}
\alias{set_crowdtangle_token}
\title{Write CT Token to system environment}
\usage{
Warning: Never store your API token in plain text within public or semi-public environments, or print it in public scripts.
}
\arguments{
\item{option}{\code{enter} if token shall be entered in a prompted RStudio window, and \code{txt} if the token is provided as a txt file.}

\item{tokenpath}{Only required if \code{option == 'txt'}. Path where txt file with token is stored.}

\item{tokenname}{Name under which token shall be saved in system environment. Defaults to "CT_TOKEN".}
}
\value{
A string with a CT API Token.
}
\description{
Allows to write a CT API token to the
system environment.
}
\examples{
# Set CT Token (Once per session)
#set_crowdtangle_token(option = "enter")
}
